Remove mypy check for unused ignore comments
There is a bug in mypy and it doesn't know what it's doing. Removing the  comment
will make mypy complain about a wrong type (even though the type is correct). Adding the  make
mypy complain about an unused ignore comment.... not sure what to do, so I just removed that warning'